Rename root namespace and add reload tool
The reload method is meant for testing/development, and not asa part of them library itself. However the root namespace ahd to be renamed from SkippyLib as the reload method otherwise collided with the reload method of the community lib I'm working on.

One day Skippy will hopefully support custom root namespace for libraries.